Move Calculator::Weight from models/open_food_network/calculator to models/calculator
