+++ config/application.rb
@@ -55,7 +55,7 @@ module Openfoodnetwork
 
     # Register Spree calculators
     initializer 'spree.register.calculators' do |app|
-      app.config.spree.calculators.shipping_methods << OpenFoodNetwork::Calculator::Weight
+      app.config.spree.calculators.shipping_methods << Calculator::Weight
       app.config.spree.calculators.add_class('enterprise_fees')
       config.spree.calculators.enterprise_fees = [
         Calculator::FlatPercentPerItem,
@@ -63,7 +63,7 @@ module Openfoodnetwork
         Spree::Calculator::FlexiRate,
         Spree::Calculator::PerItem,
         Spree::Calculator::PriceSack,
-        OpenFoodNetwork::Calculator::Weight
+        Calculator::Weight
       ]
       app.config.spree.calculators.add_class('payment_methods')
       config.spree.calculators.payment_methods = [
