+++ spec/factories.rb
@@ -275,7 +275,7 @@ FactoryBot.define do
     enterprise_role 'distributor'
   end
 
-  factory :weight_calculator, :class => OpenFoodNetwork::Calculator::Weight do
+  factory :weight_calculator, :class => Calculator::Weight do
     after(:build)  { |c| c.set_preference(:per_kg, 0.5) }
     after(:create) { |c| c.set_preference(:per_kg, 0.5); c.save! }
   end
